+++ src/lib/kaiko-sdk-node/sdk/core/data_interval_pb.d.ts
@@ -1,0 +1,34 @@
+// package: kaikosdk
+// file: sdk/core/data_interval.proto
+
+import * as jspb from "google-protobuf";
+import * as google_protobuf_timestamp_pb from "google-protobuf/google/protobuf/timestamp_pb";
+
+export class DataInterval extends jspb.Message {
+  hasStartTime(): boolean;
+  clearStartTime(): void;
+  getStartTime(): google_protobuf_timestamp_pb.Timestamp | undefined;
+  setStartTime(value?: google_protobuf_timestamp_pb.Timestamp): void;
+
+  hasEndTime(): boolean;
+  clearEndTime(): void;
+  getEndTime(): google_protobuf_timestamp_pb.Timestamp | undefined;
+  setEndTime(value?: google_protobuf_timestamp_pb.Timestamp): void;
+
+  serializeBinary(): Uint8Array;
+  toObject(includeInstance?: boolean): DataInterval.AsObject;
+  static toObject(includeInstance: boolean, msg: DataInterval): DataInterval.AsObject;
+  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
+  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
+  static serializeBinaryToWriter(message: DataInterval, writer: jspb.BinaryWriter): void;
+  static deserializeBinary(bytes: Uint8Array): DataInterval;
+  static deserializeBinaryFromReader(message: DataInterval, reader: jspb.BinaryReader): DataInterval;
+}
+
+export namespace DataInterval {
+  export type AsObject = {
+    startTime?: google_protobuf_timestamp_pb.Timestamp.AsObject,
+    endTime?: google_protobuf_timestamp_pb.Timestamp.AsObject,
+  }
+}
+