+++ src/lib/kaiko-sdk-node/sdk/stream/market_update_v1/commodity_pb.d.ts
@@ -1,0 +1,14 @@
+// package: kaikosdk
+// file: sdk/stream/market_update_v1/commodity.proto
+
+import * as jspb from "google-protobuf";
+
+export interface StreamMarketUpdateCommodityMap {
+  SMUC_UNKNOWN: 0;
+  SMUC_TRADE: 1;
+  SMUC_TOP_OF_BOOK: 2;
+  SMUC_FULL_ORDER_BOOK: 3;
+}
+
+export const StreamMarketUpdateCommodity: StreamMarketUpdateCommodityMap;
+