+++ src/services/fetchers/index.ts
@@ -13,6 +13,7 @@ import CoingeckoPriceFetcher from './CoingeckoPriceFetcher';
 import BEACPIAverageFetcher from './BEACPIAverageFetcher';
 import OnChainDataFetcher from './OnChainDataFetcher';
 import KaikoSpotPriceFetcher from './KaikoSpotPriceFetcher';
+import KaikoPriceStreamFetcher from './KaikoPriceStreamFetcher';
 
 export {
   CryptoComparePriceWSFetcher,
@@ -29,5 +30,6 @@ export {
   CoinmarketcapHistoDayFetcher,
   BEACPIAverageFetcher,
   OnChainDataFetcher,
-  KaikoSpotPriceFetcher
+  KaikoSpotPriceFetcher,
+  KaikoPriceStreamFetcher,
 };